Pillet Creek
Pillet Creek is a 2d horror game that immerses players in a dark and suspenseful environment where survival depends on wit and strategy. The game challenges players to navigate through eerie surroundings using only a flashlight, heightening the sense of urgency and fear. With limited visibility and unknown dangers lurking in the shadows, Pillet Creek aims to create a thrilling experience that keeps players on edge.
HOW TO PLAY
   - A Key: Move Left.
   - D Key: Move Right.
   - F Key: Interact with objects.
   - Left Click: Turn the flashlight on/off.
   - Right Click: Activate the flashlight's flash.